Move to a new registers file
