+++ agb/src/display/tiled.rs
@@ -1,5 +1,6 @@
 mod affine_background;
 mod infinite_scrolled_map;
+mod registers;
 mod regular_background;
 mod vram_manager;
 
@@ -13,6 +14,8 @@ pub use infinite_scrolled_map::{InfiniteScrolledMap, PartialUpdateStatus};
 pub use regular_background::{RegularBackgroundSize, RegularBackgroundTiles};
 pub use vram_manager::{DynamicTile, TileFormat, TileIndex, TileSet, VRAM_MANAGER};
 
+pub(crate) use registers::*;
+
 use bilge::prelude::*;
 
 use crate::{
@@ -22,113 +25,13 @@ use crate::{
     memory_mapped::MemoryMapped,
 };
 
-use super::{Priority, DISPLAY_CONTROL};
+use super::DISPLAY_CONTROL;
 
 #[derive(Clone, Copy, PartialEq, Eq)]
 pub struct BackgroundId(pub(crate) u8);
 
 #[derive(Clone, Copy, PartialEq, Eq)]
 pub struct AffineBackgroundId(pub(crate) u8);
-
-#[bitsize(16)]
-#[derive(FromBits)]
-struct DisplayControlRegister {
-    video_mode: u3,
-    _reserved: u1,
-    _display_frame_select: u1,
-    hblank_interval_free: bool,
-    obj_character_mapping: bool,
-    forced_blank: bool,
-    enabled_backgrounds: u4,
-    obj_display: bool,
-    window0_display: bool,
-    window1_display: bool,
-    obj_window_display: bool,
-}
-
-#[bitsize(1)]
-#[derive(Clone, Copy, FromBits, Default)]
-enum BackgroundControlTileFormat {
-    #[default]
-    FourBpp = 0,
-    EightBpp = 1,
-}
-
-impl From<TileFormat> for BackgroundControlTileFormat {
-    fn from(value: TileFormat) -> Self {
-        match value {
-            TileFormat::FourBpp => Self::FourBpp,
-            TileFormat::EightBpp => Self::EightBpp,
-        }
-    }
-}
-
-#[bitsize(1)]
-#[derive(Clone, Copy, FromBits, Default)]
-enum BackgroundControlAffineOverflowBehaviour {
-    #[default]
-    Transparent = 0,
-    Wraparound = 1,
-}
-
-impl From<AffineBackgroundWrapBehaviour> for BackgroundControlAffineOverflowBehaviour {
-    fn from(value: AffineBackgroundWrapBehaviour) -> Self {
-        match value {
-            AffineBackgroundWrapBehaviour::NoWrap => Self::Transparent,
-            AffineBackgroundWrapBehaviour::Wrap => Self::Wraparound,
-        }
-    }
-}
-
-#[bitsize(2)]
-#[derive(Clone, Copy, FromBits, Default)]
-struct BackgroundControlScreenSize(u2);
-
-impl From<RegularBackgroundSize> for BackgroundControlScreenSize {
-    fn from(value: RegularBackgroundSize) -> Self {
-        Self::new(u2::new(value as u8))
-    }
-}
-
-impl From<AffineBackgroundSize> for BackgroundControlScreenSize {
-    fn from(value: AffineBackgroundSize) -> Self {
-        Self::new(u2::new(value as u8))
-    }
-}
-
-#[bitsize(2)]
-#[derive(Clone, Copy, FromBits, Default)]
-enum BackgroundControlPriority {
-    #[default]
-    P0,
-    P1,
-    P2,
-    P3,
-}
-
-impl From<Priority> for BackgroundControlPriority {
-    fn from(value: Priority) -> Self {
-        match value {
-            Priority::P0 => Self::P0,
-            Priority::P1 => Self::P1,
-            Priority::P2 => Self::P2,
-            Priority::P3 => Self::P3,
-        }
-    }
-}
-
-#[bitsize(16)]
-#[derive(Clone, Copy, FromBits, Default)]
-struct BackgroundControlRegister {
-    priority: BackgroundControlPriority,
-    char_base_block: u2,
-    _zero: u2,
-    mosaic: bool,
-    tile_format: BackgroundControlTileFormat,
-    screen_base_block: u5,
-    overflow_behaviour: BackgroundControlAffineOverflowBehaviour,
-    screen_size: BackgroundControlScreenSize,
-}
 
 impl BackgroundId {
     #[must_use]
